fix: main finish

+++ src/domain/useCases/car/deleteCar.ts
@@ -1,4 +1,5 @@
 import { CarRepository, DeleteCar } from '@/domain/contracts';
+import { NoDataFoundError } from '@/domain/entities';
 import { formatDateTime } from '@/domain/entities/helpers';
 
 export type DeleteCar = (params: { _id: string }) => Promise<DeleteCar.Output>;
@@ -8,10 +9,16 @@ type Setup = (carRepo: CarRepository) => DeleteCar;
 export const setupDeleteCar: Setup = (carRepo) => async (params) => {
   const { _id } = params;
 
-  const car = await carRepo.delete({
+  const car = await carRepo.getById({
+    _id,
+  });
+
+  if (!car) throw new NoDataFoundError('Nenhum modelo encontrado!');
+
+  const carDeleted = await carRepo.delete({
     _id,
     deletedAt: formatDateTime(new Date()),
   });
 
-  return car;
+  return carDeleted;
 };
